Split subscription types
